refactor(platform): add platform module and flatten host firmware ioctls
Introduce platform as the public /dev/sev API and move ioctl layouts
into internal firmware::host. Remove the firmware/linux indirection and
add cpuid for host generation detection.
