--- src/firmware/linux/mod.rs
@@ -1,5 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-
-pub mod host;
-
-pub(crate) const _4K_PAGE: usize = 4096;
